move id column to NormalTable Columns
